admin dashboard update

+++ main/auth/delete.php
@@ -1,7 +1,16 @@
 <?php
+include('./auth.php');
 include("../utils/dbConnection.php");
-$id = $_GET['id'];
+if (!$authenticated) {
+    header("Location: ./login.php");
+} else {
+    if (!$admin) {
+        header("Location: ./unauth.php");
+    } else {
+        $id = $_GET['id'];
 
-$sql = "DELETE FROM users WHERE id = '$id'";
-$result = mysqli_query($conn, $sql);
-if ($result) header("Location: updateUser.php");
+        $sql = "DELETE FROM users WHERE id = '$id'";
+        $result = mysqli_query($conn, $sql);
+        if ($result) header("Location: updateUser.php");
+    }
+}
